requirements.txt
openai==1.54.0
httpx==0.27.0
pandas==2.2.0
numpy==1.26.0
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn==0.25.0
pytest==7.4.0
pytest-asyncio==0.23.0
requests==2.31.0
pydantic==2.5.0